Added UMKA_FFI feature flag & refactored ffi implementation Use umka `Storage` for allocating types Cache ffi_type structs for reuse
