+++ .clangd
@@ -1,4 +1,4 @@
 CompileFlags:
-  Add: [-DUMKA_BUILD -DUMKA_EXT_LIBS -I../libffi/build/include]
+  Add: [-DUMKA_BUILD -DUMKA_EXT_LIBS -DUMKA_FFI -I../libffi/build/include]
 InlayHints:
   Enabled: No
