+++ src/umka_ident.c
@@ -260,14 +260,17 @@ Ident *identAddBuiltinFunc(Idents *idents, const Modules *modules, const Blocks 
     return ident;
 }
 
+
+#ifdef UMKA_FFI
 Ident *identAddExternFunc(Idents *idents, const Modules *modules, const Blocks *blocks, const char *name, const Type *type, bool exported, Const constant)
 {
-    Ident *ident = identAdd(idents, modules, blocks, IDENT_EXTERN_FN, name, type, false);
+    Ident *ident = identAdd(idents, modules, blocks, IDENT_FFI_FN, name, type, false);
     ident->exported = exported;
-    ident->_extern = true;
+    ident->ffi = true;
     ident->constant = constant;
     return ident;
 }
+#endif // UMKA_FFI
 
 
 Ident *identAddModule(Idents *idents, const Modules *modules, const Blocks *blocks, const char *name, const Type *type, int moduleVal)
