+++ src/umka_lexer.c
@@ -36,8 +36,10 @@ static const char *spelling [] =
     "switch",
     "type",
     "var",
+#ifdef UMKA_FFI
+    "ffi",
+#endif
     "weak",
-    "extern",
 
     // Operators
     "+",
@@ -103,7 +105,7 @@ static const char *spelling [] =
 
 enum
 {
-    NUM_KEYWORDS = TOK_EXTERN - TOK_BREAK + 1
+    NUM_KEYWORDS = TOK_WEAK - TOK_BREAK + 1
 };
