+++ src/umka_types.h
@@ -38,7 +38,6 @@ typedef enum
     TYPE_CLOSURE,
     TYPE_FIBER,         // Pointer of a special kind
     TYPE_FN,
-    TYPE_EXTERNFN
 } TypeKind;
